Fix /nick spam with our away presence: send /nick only if nick changed
88fe1783bfd58ea0dca9e9e507a046e5c982b779 lost the "nick has actually
changed" check before sending /nick to IRC.

This reintroduces is it.

This issue became apparent with biboumi connecting to ZNC, and with
multiple xmpp clients for the same JID, resulting in a lot of presence
stanza forwarded by biboumi as a lot of /nick.
Result: libera.chat didn't like it and responded regularly with:
> lead.libera.chat: MyNick: ZNC is already trying to get this nickname